% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.interval_calculations}
\alias{plot.interval_calculations}
\title{Produce a plot for a interval_calculations object}
\usage{
\method{plot}{interval_calculations}(x, ...)
}
\arguments{
\item{x}{\code{interval_calculations} object containing the summary calculated by \code{interval}}

\item{...}{Arguments to be passed to methods}
}
\value{
object of class \code{ggplot} that contains the graphic
}
\description{
Produce a plot for a interval_calculations object
}
\author{
Trent Henderson
}
