% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thefttools.R
\docType{package}
\name{thefttools}
\alias{thefttools}
\alias{thefttools-package}
\title{Tools for Analysing and Interpreting Time Series Features}
\description{
Tools for Analysing and Interpreting Time Series Features
}
